mm/mm_heap : Split and Rearragne files about heapinfo
There was one file(mm_heapinfo.c) about heapinfo, but there were many functions in that file.
So split the files which have related functions.
mm_heapinfo.c, mm_heapinfo_group.c, mm_heapinfo_utils.c

And in sched_releasetcb.c, there was a function which related to heapinfo. So move that into the mm_heapinfo_utils.c
