+++ os/kernel/sched/sched_releasetcb.c
@@ -71,24 +71,6 @@
 #endif
 #ifdef CONFIG_DEBUG_MM_HEAPINFO
 #include <tinyara/mm/mm.h>
-
-/************************************************************************
- * Name: heapinfo_dealloc_tcbinfo
- *
- * Description:  Free the heapinfo tcb info which tcb is terminated.
- ************************************************************************/
-static void heapinfo_dealloc_tcbinfo(void *address, pid_t pid)
-{
-	pid_t hash_pid;
-	struct mm_heap_s *heap = mm_get_heap(address);
-	hash_pid = PIDHASH(pid);
-	if (heap && heap->alloc_list[hash_pid].pid == pid) {
-		heap->alloc_list[hash_pid].pid = HEAPINFO_INIT_INFO;
-		heap->alloc_list[hash_pid].curr_alloc_size = 0;
-		heap->alloc_list[hash_pid].peak_alloc_size = 0;
-		heap->alloc_list[hash_pid].num_alloc_free = 0;
-	}
-}
 #endif
 
 /************************************************************************
