+++ os/mm/mm_heap/mm_heapinfo.c
@@ -55,26 +55,14 @@
 #include <tinyara/config.h>
 #include <tinyara/sched.h>
 #include <tinyara/mm/mm.h>
-#include <tinyara/arch.h>
-#include <unistd.h>
 #include <stdint.h>
 #include <stdio.h>
-#ifdef CONFIG_HEAPINFO_USER_GROUP
-#include <string.h>
-#include <tinyara/mm/heapinfo_internal.h>
-#endif
 
 /****************************************************************************
  * Pre-processor Definitions
  ****************************************************************************/
 #define HEAPINFO_INT INT16_MAX
 #define HEAPINFO_NONSCHED (INT16_MAX - 1)
-
-#ifdef CONFIG_HEAPINFO_USER_GROUP
-int max_group;
-struct heapinfo_group_s heapinfo_group[HEAPINFO_USER_GROUP_NUM];
-struct heapinfo_group_info_s group_info[HEAPINFO_THREAD_NUM];
-#endif
 
 #if CONFIG_KMM_NHEAPS > 1
 heapinfo_total_info_t total_info;
@@ -276,255 +264,5 @@ void heapinfo_parse(FAR struct mm_heap_s *heap, int mode, pid_t pid)
 	}
 
 	return;
-}
-/****************************************************************************
- * Name: heapinfo_update_group
- *
- * Description:
- * Update Peak heap size for Group
- ****************************************************************************/
-#ifdef CONFIG_HEAPINFO_USER_GROUP
-static void heapinfo_update_group(mmsize_t size, pid_t pid)
-{
-	int check_idx;
-	int group_num;
-	int stack_pid;
-	for (check_idx = 0; check_idx < HEAPINFO_THREAD_NUM; check_idx++) {
-		if (pid == group_info[check_idx].pid) {
-			group_num = group_info[check_idx].group;
-			heapinfo_group[group_num].curr_size += size;
-			/* Update peak size */
-			if (heapinfo_group[group_num].curr_size > heapinfo_group[group_num].peak_size) {
-				heapinfo_group[group_num].peak_size = heapinfo_group[group_num].curr_size;
-				/* calculate the summation of stacks */
-				heapinfo_group[group_num].stack_size = 0;
-				for (stack_pid = 0; stack_pid < HEAPINFO_THREAD_NUM; stack_pid++) {
-					if (group_info[stack_pid].pid != -1 && group_info[stack_pid].group == group_num) {
-						heapinfo_group[group_num].stack_size += group_info[stack_pid].stack_size;
-					}
-				}
-				heapinfo_group[group_num].heap_size = heapinfo_group[group_num].peak_size - heapinfo_group[group_num].stack_size;
-			}
-			break;
-		}
-	}
-}
-#endif
-/****************************************************************************
- * Name: heapinfo_add_size
- *
- * Description:
- * Add the allocated size in tcb
- ****************************************************************************/
-void heapinfo_add_size(struct mm_heap_s *heap, pid_t pid, mmsize_t size)
-{
-	pid_t hash_pid;
-
-	hash_pid = PIDHASH(pid);
-	if (heap->alloc_list[hash_pid].pid == HEAPINFO_INIT_INFO || heap->alloc_list[hash_pid].pid == pid) {
-			heap->alloc_list[hash_pid].pid = pid;
-			heap->alloc_list[hash_pid].curr_alloc_size += size;
-			if (heap->alloc_list[hash_pid].curr_alloc_size > heap->alloc_list[hash_pid].peak_alloc_size) {
-				heap->alloc_list[hash_pid].peak_alloc_size = heap->alloc_list[hash_pid].curr_alloc_size;
-			}
-			heap->alloc_list[hash_pid].num_alloc_free++;
-	}
-}
-
-/****************************************************************************
- * Name: heapinfo_subtract_size
- *
- * Description:
- * Subtract the allocated size in tcb
- ****************************************************************************/
-void heapinfo_subtract_size(struct mm_heap_s *heap, pid_t pid, mmsize_t size)
-{
-	pid_t hash_pid;
-
-	hash_pid = PIDHASH(pid);
-	if (heap->alloc_list[hash_pid].pid == pid) {
-			heap->alloc_list[hash_pid].curr_alloc_size -= size;
-			heap->alloc_list[hash_pid].num_alloc_free--;
-	}
-}
-
-/****************************************************************************
- * Name: heapinfo_update_total_size
- *
- * Description:
- * Calculate the total allocated size and update the peak allocated size for heap
- ****************************************************************************/
-void heapinfo_update_total_size(struct mm_heap_s *heap, mmsize_t size, pid_t pid)
-{
-	heap->total_alloc_size += size;
-	if (heap->total_alloc_size > heap->peak_alloc_size) {
-		heap->peak_alloc_size = heap->total_alloc_size;
-	}
-#if CONFIG_KMM_NHEAPS > 1
-	total_info.cur_alloc_size += size;
-	if (total_info.cur_alloc_size > total_info.peak_alloc_size) {
-		total_info.peak_alloc_size = total_info.cur_alloc_size;
-	}
-#endif
-#ifdef CONFIG_HEAPINFO_USER_GROUP
-	heapinfo_update_group(size, pid);
-#endif
-}
-/****************************************************************************
- * Name: heapinfo_update_node
- *
- * Description:
- * Adds pid and malloc caller return address to mem chunk
- ****************************************************************************/
-void heapinfo_update_node(FAR struct mm_allocnode_s *node, mmaddress_t caller_retaddr)
-{
-	node->alloc_call_addr = caller_retaddr;
-	node->reserved = 0;
-	node->pid = getpid();
-	return;
-}
-
-/****************************************************************************
- * Name: heapinfo_exclude_stacksize
- *
- * Description:
- * when create a stack, subtract the stacksize from parent
- ****************************************************************************/
-void heapinfo_exclude_stacksize(void *stack_ptr)
-{
-	struct mm_allocnode_s *node;
-	pid_t hash_pid;
-	struct mm_heap_s *heap = mm_get_heap(stack_ptr);
-	if (heap == NULL) {
-		return;
-	}
-	node = (struct mm_allocnode_s *)(stack_ptr - SIZEOF_MM_ALLOCNODE);
-	hash_pid = PIDHASH(node->pid);
-	heap->alloc_list[hash_pid].curr_alloc_size -= node->size;
-#ifdef CONFIG_HEAPINFO_USER_GROUP
-	int check_idx;
-	int group_num;
-
-	for (check_idx = 0; check_idx <= max_group; check_idx++) {
-		if (node->pid == group_info[check_idx].pid) {
-			group_num = group_info[check_idx].group;
-			heapinfo_group[group_num].curr_size -= node->size;
-			break;
-		}
-	}
-#endif
-}
-
-#ifdef CONFIG_HEAPINFO_USER_GROUP
-/****************************************************************************
- * Name: heapinfo_update_group_info
- *
- * Description:
- * when create or release task/thread, check that the task/thread is 
- * in group list
- ****************************************************************************/
-void heapinfo_update_group_info(pid_t pid, int group, int type)
-{
-	int info_idx;
-	struct tcb_s *tcb;
-	switch (type) {
-	case HEAPINFO_INIT_INFO:
-		for (info_idx = 0; info_idx < HEAPINFO_THREAD_NUM; info_idx++) {
-			group_info[info_idx].pid = pid;
-			group_info[info_idx].group = group;
-			group_info[info_idx].stack_size = 0;
-		}
-		break;
-	case HEAPINFO_ADD_INFO:
-		if (group > max_group) {
-			max_group = group;
-		}
-		for (info_idx = 0; info_idx < HEAPINFO_THREAD_NUM; info_idx++) {
-			if (group_info[info_idx].pid <= 0) {
-				group_info[info_idx].pid = pid;
-				group_info[info_idx].group = group;
-				tcb = sched_gettcb(pid);
-				if (tcb) {
-					group_info[info_idx].stack_size = tcb->adj_stack_size;
-					heapinfo_update_group(tcb->adj_stack_size, pid);
-				}
-				break;
-			}
-		}
-		break;
-	case HEAPINFO_DEL_INFO:
-		for (info_idx = 0; info_idx < HEAPINFO_THREAD_NUM; info_idx++) {
-			if (pid == group_info[info_idx].pid) {
-				heapinfo_update_group((-1) * group_info[info_idx].stack_size, pid);
-
-				group_info[info_idx].pid = -1;
-				group_info[info_idx].group = -1;
-				group_info[info_idx].stack_size = 0;
-				break;
-			}
-		}
-		break;
-	default:
-		break;
-	}
-}
-
-/****************************************************************************
- * Name: heapinfo_check_group_list
- *
- * Description:
- * check that task/thread is in group list
- ****************************************************************************/
-void heapinfo_check_group_list(pid_t pid, char *name)
-{
-	char *thread_list = CONFIG_HEAPINFO_USER_GROUP_LIST;
-
-	int group_num = 0;
-
-	char *name_start;
-	char *name_end;
-	int name_length;
-	name_start = name_end = thread_list;
-
-	while (*name_start != '\0') {
-		if (*name_end == '/' || *name_end == ',' || *name_end == '\0') {
-			name_length = name_end - name_start;
-			if (name_length == strlen(name)) {
-				if (strncmp(name_start, name, name_end - name_start) == 0) {
-					heapinfo_update_group_info(pid, group_num, HEAPINFO_ADD_INFO);
-					break;
-				}
-			}
-			if (*name_end == '/') {
-				group_num++;
-			} else if (*name_end == '\0') {
-				name_start = name_end;
-				continue;
-			}
-			name_end++;
-			name_start = name_end;
-		} else {
-			name_end++;
-		}
-	}
-}
-#endif /* CONFIG_HEAPINFO_USER_GROUP */
-
-/****************************************************************************
- * Name: heapinfo_peak_init
- *
- * Description:
- * initialize the peak allocation size in heap
- ****************************************************************************/
-void heapinfo_peak_init(struct mm_heap_s *heap)
-{
-	int tcb_idx;
-	int heap_idx;
-
-	for (heap_idx = 0; heap_idx < CONFIG_KMM_NHEAPS; heap_idx++) {
-		for (tcb_idx = 0; tcb_idx < CONFIG_MAX_TASKS; tcb_idx++) {
-			heap[heap_idx].alloc_list[tcb_idx].peak_alloc_size = 0;
-		}
-	}
 }
 #endif
